Fixed app path for deployment to dev (#9)
changed package.json to set an env variable flag when running in local mode with moock API (VUE_APP_MOCK_ENV === 'yes').  WHen it is set to yes, the publicPath is '/' and the API url is set to localhost.  Otherwise, the VUE_APP_PATH and VUE_APP_API_URL are used.